www/requirements.txt
Flask
datetime
uuid
Flask-SQLAlchemy
PyJWT
Flask-Restful
Flask-Sqlalchemy
Flask-Migrate
Flask-Marshmallow
Marshmallow-Sqlalchemy
PyJWT
flask-jwt-extended
Flask-Cors
python-slugify
